+++ src/SMAPI/Events/NpcListChangedEventArgs.cs
@@ -20,6 +20,9 @@ namespace StardewModdingAPI.Events
         /// <summary>The NPCs removed from the location.</summary>
         public IEnumerable<NPC> Removed { get; }
 
+        /// <summary>Whether this is the location containing the local player.</summary>
+        public bool IsCurrentLocation => object.ReferenceEquals(this.Location, Game1.player?.currentLocation);
+
 
         /*********
         ** Public methods